requirements.txt
pyautogui
pywin32
selenium